Get x64 running with the new asm model.
The generic JIT core issues simple RISC-like instructions to the asm
backends, and the backend can coalesce them where it makes sense in
the backend instruction set.

x64 coalesces a lot; ARM64, when updated, is expected to do less
coalescing.

The optimizer will need re-doing; it's thoroughly bricked by this
change.

+++ jit_opcode.h
@@ -22,7 +22,7 @@ struct jit_opcode_details {
   int32_t max_6502_addr;
 
   /* Partially dynamic details that may be changed by optimization. */
-  uint8_t num_uops;
+  uint32_t num_uops;
   struct asm_uop uops[k_max_uops_per_opcode];
   struct asm_uop* p_prefix_uop;
   struct asm_uop* p_postfix_uop;
